pr #1365. Added some tests for translation from tl schedule to ad. Also removed old schedule transformation and associated tests.
